// src/app/site-premiando/sorteios/models/full-sorteio.ts
import { SafeHtml } from "@angular/platform-browser";

export class FullSorteio{
  
  dataEncerramento: string;
  dataSorteio: string;
  descricao: string;
  exibirData: boolean;
  id: number;
  imagem: string;
  keywords: string;
  numeroFinal: number;
  numeroInicial: number;
  ocultar: boolean;
  rifinha: boolean;
  status: number;
  texto: string;
  adicional1: string;
  adicional2: string;
  textoHTML: SafeHtml;
  titulo: string;
  valor: number;
  whatsapp: string;
}